requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
PyJWT==2.10.1
pyotp==2.9.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
